Move original tests into original folder
