+++ code/test/original/basic.js
@@ -1,11 +1,11 @@
-const TestUtil = require('./util')
+const TestUtil = require('../util')
 const unit = require('ethjs-unit')
-const ONEUtil = require('../lib/util')
-const ONEDebugger = require('../lib/debug')
-const ONE = require('../lib/onewallet')
-const ONEConstants = require('../lib/constants')
-const Flow = require('../lib/api/flow')
-const ONEWallet = require('../lib/onewallet')
+const ONEUtil = require('../../lib/util')
+const ONEDebugger = require('../../lib/debug')
+const ONE = require('../../lib/onewallet')
+const ONEConstants = require('../../lib/constants')
+const Flow = require('../../lib/api/flow')
+const ONEWallet = require('../../lib/onewallet')
 const BN = require('bn.js')
 
 const INTERVAL = 30000
