+++ code/test/original/spendLimit.js
@@ -1,9 +1,9 @@
-const TestUtil = require('./util')
+const TestUtil = require('../util')
 const unit = require('ethjs-unit')
-const ONEUtil = require('../lib/util')
-const ONEDebugger = require('../lib/debug')
-const ONEWallet = require('../lib/onewallet')
-const ONEConstants = require('../lib/constants')
+const ONEUtil = require('../../lib/util')
+const ONEDebugger = require('../../lib/debug')
+const ONEWallet = require('../../lib/onewallet')
+const ONEConstants = require('../../lib/constants')
 const BN = require('bn.js')
 const INTERVAL = 30000
 const INTERVAL6 = INTERVAL * 6
